Rename Inputs to FileInputs. Separate Fastdoop-specific input logic from general input logic.
