+++ src/test/scala/com/jnpersson/slacken/Testing.scala
@@ -19,8 +19,9 @@
 package com.jnpersson.slacken
 
 import com.jnpersson.kmers
+import com.jnpersson.kmers.input.FileInputs
 import com.jnpersson.kmers.minimizer._
-import com.jnpersson.kmers.{AnyMinSplitter, IndexParams, Inputs, TestGenerators, Testing => TTesting}
+import com.jnpersson.kmers.{AnyMinSplitter, IndexParams, TestGenerators, Testing => TTesting}
 import com.jnpersson.slacken.Taxonomy.{NONE, ROOT, Rank, Root}
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.scalacheck.{Gen, Shrink}
@@ -170,7 +171,7 @@ object TestData {
   val numberOf35Mers = Map(526997 -> 2902850, 455631 -> 3565872, 9606 -> 639784)
 
   def inputs(k: Int)(implicit spark: SparkSession) =
-    new Inputs(List("testData/slacken/slacken_tinydata.fna"), k, 10000000)
+    new FileInputs(List("testData/slacken/slacken_tinydata.fna"), k, 10000000)
 
   def library(k: Int)(implicit spark: SparkSession): GenomeLibrary =
     GenomeLibrary(inputs(k), "testData/slacken/seqid2taxid.map")
